test(datepicker): update/add tests for new structure
affects: @fremtind/jkl-datepicker-react
